Refactor ANSI attribute types into extensions
